% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableDietaryData.R
\name{availableDietaryData}
\alias{availableDietaryData}
\title{Fetch a table of the different NHANES dietary datasets available for a given survey cycle.}
\usage{
availableDietaryData(start.year = "2017")
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}
}
\value{
A data.frame containing the names of the available dietary datasets for the given year and their abbreviated names.
}
\description{
This function generates a table of the available NHANES dietary datasets and their abbreviated names (e.g. "Dietary Supplement Database - Ingredient Information" and "DSII" for 2017-2018). The desired abbreviated values can then be fed into \code{\link{fetchDietaryData}}, which will return the dietary data itself.
}
\examples{
availableDietaryData(start.year = "2013")
}
\seealso{
\code{\link{fetchDietaryData}}
}
\author{
Jack Leary
}
