% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableExaminationData.R
\name{availableExaminationData}
\alias{availableExaminationData}
\title{Fetch a table of the different NHANES physician examination datasets available for a given year.}
\usage{
availableExaminationData(start.year = "2017")
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}
}
\value{
A data.frame containing the names of the available examination datasets for the given year and their abbreviated names.
}
\description{
This function generates a table of the available NHANES physician examination datasets and their abbreviated names e.g., "Blood Pressure" and "BPX_J" for 2017-2018. The desired abbreviated values can then be fed into \code{\link{fetchExaminationData}}, which will return the examination data itself.
}
\examples{
availableExaminationData(start.year = "2013")
}
\seealso{
\code{\link{fetchExaminationData}}
}
\author{
Jack Leary
}
