% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableLabData.R
\name{availableLabData}
\alias{availableLabData}
\title{Fetch a table of the different NHANES laboratory datasets available for a given survey cycle.}
\usage{
availableLabData(start.year = "2017")
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}
}
\value{
A data.frame containing the names of the available lab datasets for the given year and their abbreviated names.
}
\description{
This function generates a table of the available NHANES laboratory datasets and their abbreviated names (e.g. "Cholesterol - Total" and "TCHOL_J" for 2017-2018). The desired abbreviated values can then be fed into \code{\link{fetchLabData}}, which will return the lab data itself.
}
\examples{
availableLabData(start.year = "2013")
}
\seealso{
\code{\link{fetchLabData}}
}
\author{
Jack Leary
}
