% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSurveyDesign.R
\name{bootSurveyDesign}
\alias{bootSurveyDesign}
\title{Generate replicate weights for complex survey designs.}
\usage{
bootSurveyDesign(
  survey.dat = NULL,
  id.var = NULL,
  strata.var = NULL,
  weight.var = NULL,
  nested.clusters = TRUE,
  boot.type = "bootstrap",
  n.boot = 1000
)
}
\arguments{
\item{survey.dat}{A data.frame containing survey data. Defaults to NULL.}

\item{id.var}{A string containing the name of the variable that denotes clusters of primary sampling units. Defaults to NULL.}

\item{strata.var}{A string containing the name of the variable that stratifies the PSUs. Defaults to NULL.}

\item{weight.var}{A string containing the name of the variable containing sampling weights. Defaults to NULL.}

\item{nested.clusters}{Are the clusters nested within strata? Defaults to TRUE}

\item{boot.type}{A string specifying the resampling method. Passed to the \code{type} argument of \code{\link[survey]{as.svrepdesign}}. Defaults to "bootstrap".}

\item{n.boot}{The number of replications to perform. Defaults to 1000.}
}
\value{
An object of class \code{svyrep.design} containing replicate weights.
}
\description{
This function creates a \code{survey.design} object using the \code{survey} package, then uses resampling to generate sample weights.
}
\examples{
\dontrun{bootSurveyDesign(survey.dat = nhanes_data, id.var = "CLUST_ID", strata.var = "STRATUM", weight.var = "SAMPLE_WEIGHT")}
}
\seealso{
\code{\link[survey]{svydesign}} \code{\link[survey]{as.svrepdesign}}
}
\author{
Jack Leary
}
