% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDemographicData.R
\name{fetchDemographicData}
\alias{fetchDemographicData}
\title{Fetch NHANES demographics data for a given year.}
\usage{
fetchDemographicData(start.year = "2017")
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}
}
\value{
A data.frame containing the requested survey data.
}
\description{
This function pulls demographic survey data from the CDC's NHANES survey from a user-provided year.
}
\examples{
fetchDemographicData(start.year = "2017")
}
\author{
Jack Leary
}
