% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDietaryData.R
\name{fetchDietaryData}
\alias{fetchDietaryData}
\title{Fetch NHANES dietary data for a given year.}
\usage{
fetchDietaryData(start.year = "2017", diet.dataset.list = NULL)
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}

\item{diet.dataset.list}{A character vector containing the names of specific dietary datasets the user would like to analyze. If none are provided, all datasets are returned. Defaults to NULL.}
}
\value{
A named list containing the requested dietary datasets.
}
\description{
This function pulls various types of supplementary dietary data from the CDC's NHANES survey for a user-provided year.
}
\examples{
fetchDietaryData(start.year = "2017", diet.dataset.list = c("DR1IFF_J", "DR2IFF_J"))
lapply(c("2015", "2017"), fetchDietaryData)
}
\seealso{
\code{\link{availableDietaryData}}
}
\author{
Jack Leary
}
