% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLabData.R
\name{fetchLabData}
\alias{fetchLabData}
\title{Fetch NHANES lab data for a given survey cycle.}
\usage{
fetchLabData(start.year = "2017", lab.dataset.list = NULL)
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}

\item{lab.dataset.list}{A character vector containing the names of specific lab datasets the user would like to analyze. If none are provided, all lab datasets are returned. Defaults to NULL.}
}
\value{
A named list containing the requested lab datasets.
}
\description{
This function pulls various types of laboratory data from the CDC's NHANES survey for a user-provided year.
}
\examples{
fetchLabData(start.year = "2017", lab.dataset.list = c("TCHOL_J", "PBCD_J"))
}
\seealso{
\code{\link{availableLabData}}
}
\author{
Jack Leary
}
