% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchQuestionnaires.R
\name{fetchQuestionnaires}
\alias{fetchQuestionnaires}
\title{Fetch NHANES questionnaire data for a given year.}
\usage{
fetchQuestionnaires(start.year = "2017", survey.list = NULL)
}
\arguments{
\item{start.year}{The year for which the user desires data. Must not be later than 2017 (the last year currently available). Defaults to "2017".}

\item{survey.list}{A character vector containing the names of specific questionnaires the user would like to analyze. If none are provided, all questionnaires are returned. Defaults to NULL.}
}
\value{
A named list containing the requested questionnaire data.
}
\description{
This function pulls more granular questionnaire data from the CDC's NHANES survey from a user-provided year.
}
\examples{
fetchQuestionnaires(start.year = "2017")
lapply(c("2015", "2017"), fetchQuestionnaires)
}
\seealso{
viewAllQuestionnaires
}
