% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseQuantReg.R
\name{summariseQuantReg}
\alias{summariseQuantReg}
\title{Get summary stats for quantile regression models of survey data.}
\usage{
summariseQuantReg(quant.reg.fit = NULL, boot.weights = NULL, alpha = 0.05)
}
\arguments{
\item{quant.reg.fit}{The bootstrapped coefficients from a quantile regression model. Defaults to NULL.}

\item{boot.weights}{Bootstrapped replicate weights as generated using \code{\link{bootSurveyDesign}}. Defaults to NULL.}

\item{alpha}{The desired Type I Error Rate. Defaults to \eqn{\alpha = 0.05}.}
}
\value{
A data.frame containing variable names, coefficients, standard errors, test statistics, and \emph{p}-values.
}
\description{
This function summarises the coefficients of quantile regression models.
}
\examples{
\dontrun{summariseQuantReg(quant.reg.fit = my_model, boot.weights = boot_sample_weights, alpha = 0.01)}
}
\seealso{
\code{\link{bootSurveyDesign}} \code{\link[quantreg]{rq}}
}
\author{
Jack Leary
}
